move: mapUtils : domain -> email 패키지 로 이동[AIR-144]
